# docs/api/student-api.openapi.yaml
openapi: 3.1.0
info:
  title: Student + Parent API (sms-student)
  version: "1.0.0-phase5-contract"
  description: >
    Forward contract for the student/parent app. Phase-5 deliverable; backend currently implements only
    Phase 0. snake_case DTOs mirror sms-student/src/services/types.ts. Two roles (student, parent).
    Session carries refresh_token (app currently discards it). See markdown twin for reconciliation notes.
servers:
  - url: "{apiBaseUrl}/v1"
    variables: { apiBaseUrl: { default: https://api.catre.io } }
security:
  - bearerAuth: []
...
paths:
...
  /homework/{id}:
    parameters: [{ $ref: "#/components/parameters/Id" }]
    get: { tags: [academics], responses: { "200": { description: ok, content: { application/json: { schema: { $ref: "#/components/schemas/Homework" } } } } } }
    patch:
      tags: [academics]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [status], properties: { status: {type: string, enum: [todo, progress, submitted, graded]} } } } } }
      responses: { "200": { description: ok, content: { application/json: { schema: { $ref: "#/components/schemas/Homework" } } } } }
  /homework/{id}/submit:
    parameters: [{ $ref: "#/components/parameters/Id" }]
    post: { tags: [academics], responses: { "200": { description: ok, content: { application/json: { schema: { $ref: "#/components/schemas/Homework" } } } } } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    Id: { name: id, in: path, required: true, schema: { type: string } }
...
  schemas:
...
    Homework:
      type: object
      properties: { id: {type: string}, assignment_id: {type: string}, title: {type: string}, subject_id: {type: string}, due_date: {type: string}, due_time: {type: string}, status: {type: string, enum: [todo, progress, submitted, graded]}, priority: {type: string, enum: [low, med, high]}, grade: {type: string} }
...